Added auto model assignment for kicad_v4
Read kicad_pcb model set in the footprints, and automatically populate
the database if a supported 3D model file is found at search path. In
other words, all you need to do is to put the STEP/IGES file with the
same name in the same location as the VRML files. The STEP/IGES model
shall have the same placement as the VRML one.